code/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
googletrans==3.1.0a0
python-dotenv==1.0.1
werkzeug==3.0.1